:construction: First Round Of Middleware Categorization
